simplify but moving Dialect::load -> Driver::load
